+++ lib/hexapdf/importer.rb
@@ -117,7 +117,7 @@ module HexaPDF
       when HexaPDF::Object
         if object.type == :Catalog || object.type == :Pages
           @mapper[object.data] = nil
-        elsif (mapped_object = @mapper[object.data]) && !mapped_object.null?
+        elsif (mapped_object = @mapper[object.data]&.__getobj__) && !mapped_object.null?
           mapped_object
         else
           obj = object.dup